+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - When an user authenticate we issue a device.login event
+- Add a contact from a discussion
 
 ### Fixed